The dalton script now modified to keep all files that may potentially be of interest to a user in postprocessing or as start for new calculations. The files are stored in the file dal_mol.tar.gz.
The script will now also allow for only one input file name, so that
"dalton hei"  corresponds to the old   "dalton hei hei".

By default, the job will now run in the scratch directory "dal_mol" instead of
the old default of "mol".

This new system must be checked/improved with respect to STEX, linear coupling
models and other calculations that use multiple resulting files. Currently not
dealt with and left as an excercise for the user. \kr\