Workaround of arithmetic overflow errors produces by gfortran associated with assignment of hexadecimal consants (typically mask variables) to INTEGER*4 variables.
Changes are set in #if defined (VAR_GFORTRAN) and passed the allcc test suite

Note that gfortran does the correct thing according to f95 standard: Hexadecimal
constants are interpreted in terms of the largest available integer KIND on the
system. E.g. the hexadecimal bit pattern z'C0000000' is, when INTEGER*8 is
supported, a positive number outside the range of INTEGER*4 numbers (where
the same bit pattern is a negative number -'z40000000').

OV