Undoing last commit. It is not good to mix logical modules, in this case the new LUCITA and the 2-electron integral transformation is sirius/*tra*.F.
The interface should be as clean as possible. 
Thus all orbital space information to the integral transformation (for
now) should ALWAYS be passed via inforb.h (nish, nash, norb, nbas).
Thus the proper way (for now) to change the integral transformation is
to set nish, nash, norb, nbas to the desired values. Note that inforb.h
is supposed to be the current relevant values of the different orbital
spaces, if you look in the SIRIUS code you will see that they change
from Hartree-Fock to CI/MCSCF.

It is desirable to implement the string selection used in DIRAC as soon
as someone has time to do it ... (thus the "for now" above)