Changed code for automatic selection of MS2 / MULTS
Moved automatic selection out in sirinp.F(SIRRDI) so that all relevant input
always have been read (in particular .ACTIVE ELECTRONS and .DETERMINANTS).

Print 2 M_S value in output and changed tests which check this line in output.

Check for consistency if both .SPIN MULTIPLICITY and .MS2 are specified.

This also fixes that some g-tensor and ZFS tests were relying on that
.DETERMINANTS meant that M_S = S.

Fix of output error in siropt.F.