New '-nb mb' option for slave WORK memory, '-M' renamed to '-mb' and some polish.
With the new -nb option it is possible to let the slaves allocate much less memory than
the master, e.g. '-mb 14000 -nb 2000' giving 14GB to master and 2GB to each slave.
'-nb' exports the new NODE_WRKMEM environment variable with the specified value.

The old -M was renamed to -mb (but -M is still understood).

Updated usage message.