Implemented MP2-NO for HSROHF with frozen singly occupied orbitals.
Should have thought of this many years ago! The correlation effects
are by far biggest for the doubly occupied orbitals. By freezing the
singly occupied orbitals the MP2 equations for the rest are easy: the
same as for the closed-shell case! In this way we can now get good
orbitals for MCSCF also for open shell cases :-) -- much cheaper than
the CISD-NO which hitherto has been the only option. I also expect that
the NO occupations will be useful for selecting active spaces, just
as in the closed shell cases.

Also some polish here and there of output and comments.
(Incl. writing ERROR instead of error, such that grep in the dalton script
catches them.)