New "ERFEXP" long-range/short-range separation
1) logical ERFEXP in gnrinf.h changed to ERFEXP(0:2)
   ERFEXP(1) is old erfgau separation
   ERFEXP(2) is Morten's new erfgau separation
   ERFEXP(0) = ERFEXP(1) .OR. ERFEXP(2)
2) input option .ERFEXP now activates new erfgau
   input option for old erfgau has been changed to .ERFGAU, in agreement
   with that this separation is called erfgau in published papers.

The old erfgau uses the gaussian term to make the long-range potential
start at zero, that is, to model a vertical separation between short-range
and long-range.

The new erfgau uses the gaussian term to make the long-range potential
start as flat as possible, that is, a constant (horisontal) long-range
potential for small distances. This should give simpler correlation
functionals because there is no correlation for a constant potential.

  Morten and Hans Joergen.