Partial bugfix of PBHT overlap diagnostic when symmetry
The "Lambda" overlap diagnostic is wrong with symmetry.
Thus bugfix fixes wrong addressing in CMO when symmetry.
Unfortunately there are more errors -- the results are wrong
for H2O in C2v symmetry :(

The new cpp option #define DFTMOS_DEBUG makes the integration
routine calculate the MO overalp matrix, easy to check if it
is a unit matrix. For H2O in C2v there are non-zero off-diagoanl
elements between symmetries 1 and 3, and between 2 and 4.
The diagonal blocks are now unit matrices, after the off-set in
CMO was fixed.