Fixed error in timing output for NEWTRA by removing the output !
Because it was printed twice, once correct and once wrong.
Have also inserted a FLSHFO in siropt.F before the final integral
transformation because I was waiting impatiently for some output in
a test of the integral transformation.
Have started a move to replace MXMA with DGEMM in sirntra.F, not finished yet.