hjaaj: corrected error in manual as reported by user:
Hi Martin

The sample input in the manual is incorrect in this case.
The information for the abacus module about how the properties are to be calculated is should be contained  in an **EACH STEP  section not in a **PROPERTIES section.
So replacing **PROPERTIES with **EACH STEP in your input, should make it work.

Hope this helps

Rasmus Faber
________________________________________
Fra: Martin Holzschuh [martin.holzschuh@uni-ulm.de]
Sendt: 3. december 2012 14:28
Til: dalton-users@kjemi.uio.no
Emne: [dalton-users] Calculation of Vibrationally averaged Spin-Spin Coupling Constants

Dear Dalton Users and Developers,

I am trying to run ZPVC on spin-spin coupling constants on our 64bit linux-cluster. So first i tried the example (H2O) from the manual but it fails with the message provided in the *log  file. Any ideas what i did wrong?

Thanks in advance, and greetings to all.

Martin Holzschuh