Tamm-Dancoff approximation to TDDFT or RPA (= CIS)
The TDA is invoked by the use of the .TDA or .CIS keywords under
**RESPONSE or *LINEAR, however, if both DODFT and CISRPA logicals
are .true., then exit with a line telling to use .TDA keyword instead.
**NOTE: NO real computational efficiency is gained over TDDFT/RPA by
using the TDA, as we only set B=0 and avoid inserting anything into
the Y matrix, i.e., a TDA calculation comes at the price of a regular
TDDFT/TDHF calculation.