Corrected problems in symmetry analysis
The symmetry identification had problems, when there were several different
sets of symmetry related atoms of a given type.
This problem manifested itself much more frequently as the sorting
order of atoms of the same type was reversed in c5f9a0af

Certain treshholds, mostly in the ROTAXS function,
was not consistent with the printed treshhold,
causing problems with "allmost symmetrical" geometries encountered when doing
numerical differentiation.