hjaaj: more intelligent .MOSTART NEWORB for SCF
Had disabled reading SCF occupation of symmetries in 2008, because you may start
from orbitals from a calculation with another total charge. This meant that
H1DIAG guess was used for initial occupation of symmetries. Now instead, we
check if number of electrons has changed, and if not we use the old occupation.