hjaaj: do not waste memory any more by copying DMAT in SIRFCK
DMAT is destroyed in some AODIRECT calculations, but in most cases
DMAT is not needed afterwards. Now it is copied in FCK2AO in sirgrad.F
instead, as this is the only place where it is needed afterwards.
It makes a difference if you e.g. want to calculate 200 fock matrices in response.

And polish in siraba.F.