Option to make plot files after orb. localization
Added keyword .ORBITAL PLOT which should be followed by
HOMO, LUMO, MOSTL, LEASTL or ALL, which will specify which
plt files will be generated. If HOMO or LUMO is used, the
file HOMO.plt or LUMO.plt will be generated such that the HOMO
or LUMO may be plotted using a plotting program. If MOSTL is used
the files MostLocal_occ.plt and MostLocal_virt.plt will be generated
corresponding to the plotting files for the most local occupied and
virtual orbitals. Similarly for LEASTL where LeastLocal_occ.plt and
LeastLocal_virt.plt will be generated. If ALL is used both the most
and least local orbital plot files will be generated (MostLocal_occ.plt,
MostLocal_virt.plt,LeastLocal_occ.plt and LeastLocal_virt.plt), but
not the HOMO.plt and LUMO.plt.

.ORBITAL PLOT will only be used if orbital localization is requested.