Modified SCF procedure:
The New default for TRILEVEL starting guess is to include the
ATOMS starting guess to the list of previous density matrices, which
is a substantial advantage when the ATOMS starting guess is a better starting point
than the TRILEVEL start. I also modified the use og the gradient norm to always use
AO gradient or OAO gradient. Before it actually used AO gradient in the first iteration
and then switch to print the OAO gradient. In addition the OAO gradient it printet was
not the completly correct one. This means that a number of testcases was changed due to
different number of SCF iterations, due to different gradnrm used for convergence test.

I also fixed the ddynam/H2O_fmd test case after talking with Vladimir.