Remove .NODIIS option from four cc3 test cases.
The -ffast-math option miscompiles ONEPR1 and HERNAI
subroutines (both in abacus/her1car.F) on some systems.
The miscompilation leads to breaking the following cc3
test cases:
	cc3_LIH1s1p_asym_pol
	cc3_LIH1s1p_pol
	cc3_LIH1s1p_tmom
	cc3_LIH1s1p_tmom_no2np1
(which use a basis of one s and one p function within
the INTGRL style of mol input).
Removing the .NODIIS option from the dal inputs for
these test cases fixes the problem.
Note, however, that THIS IS NOT A BUG FIX!!!