Using an automaticaly generated reorderings module
this has been placed in lsutil/generate_man_reord.py and depends on the
header written in lsutil/reorder_header.F90. Within some limits the
script has the capability of finding the directories itself and placing
the manual_reorderings.F90 accordingly.

building is enabled both with cmake and makefile.config.