Added quit statements which makes sense
Recommend using .ARH if singularities are encountered
for SCF optimization using ARH/davidson solver.

FOr orbital localization, this shouldn't happen due to the
previous commit, but if it happens, quit statement will ask
user to break molecular symmetry slightly. This is not a pretty
thing, but it's a sure fix if the unthinkable happens.