hjaaj: someone wanted to print the coordinates in READIN in every geometry iteration for POLEMB
This has been disabled, user has to ask specifically for that.
I am guessing what was really wanted was to print all coordinates when POLEMB
with default print level, also when more than 13 atoms. This I have now programmed.

(The coordinates in geometry optimizations are always printed at end of step, so
the printing in READIN is an unnecessary duplication, thus it should not be default.
Even for POLEMB.)