Change of basis set in cc3_LIH1s1p* tests
The series of cc3_LIH1s1p* test cases was failing
on some architectures / for some compilers / libraries.
This is due to a very strange basis (1s function and
1p function, which apparently leads to multiple
solutions for the CC3 ground state).

I thus changed this small and strange basis to
cc-pVDZ basis for all the cc3_LIH1s1p* tests, hoping i
that will fix the problem.

The names of the tests are kept intact for the moment.