General cleanup of the coupled cluster code
this comes at the expense of one more crop solver routine which calls
the debug and reference routines in cc_debug_routines.F90 where all the
reference stuff is put together. This file is there mainly for developing
purposes.

In the long run the more general ccsolver_par (which does a good job in keeping the
data local as well) should replace the ccsolver.