bugfix in GPOPEN by T. Keal
radovan: email text
"""
Finally I found a new bug in the GPOPEN subroutine gp/gptrygve.F. The problem
is when a .n000x prefix is added to the slave nodes file names there is a call
to INDEX(FILENM) - but this searches over the whole of FILENM, which is not
cleared at the start of the subroutine, and so may contain stray ".n"'s in the
undefined section. This was causing my calculations to hang on the second call
to the DALTON subroutine. I have fixed this by calling INDEX(FILENM(1:FILELN)
in each case (see attached file).
"""