clean startup and shutdown for comm procs on several slaves
it is now possible to start up additional processes runnning lsdalton
with the function
 - give_birth_to_child_process
if the slaves should be convinced to do the same, either they call the
same routine or the master bcasts GIVE_BIRTH, and the slaves follow as
ordered. PLEASE NOTE: this works only if the executable "lsdalton.x" is
in the working directory, otherwise the code quits:
For DEC-programmers: If you want to use that feature (Janus) then I
suggest to use the DECinfo%spawn_comm_proc variable which is set to
.true. with the keyword ".SPAWN_COMM_PROC" and the code should work with
and without it being set, since the comm-proc approach usually uses MPI3
features

To get rid of the child again we can simply call
 -shut_down_child_process
and to order the slaves to do the same we broadcast
SLAVES_SHUT_DOWN_CHILD.

To be able to do this a restructuring of the (not very pretty) previous
code in routine LSDALTON was made and the slaves are not killed of by
hitting a simple "stop" statement anymore. The slaves, childs and the
master call the same initialization and finalization routine now, which
makes the code much more clearer as everything that happens during
startup and shutdown is controled by the same routines. Thomas nice
memory-info printing routine is now an own routine and NOT part of the
lsmpi_finalize anymore (as it should be in my opinion).

If you have questions, comments or feel offended in any way, feel free
to write me