hjaaj: bugfix for frozen coordinates in geometry optimization
When nuclei are frozen it is wrong to project out translation and rotation.
New PRJTRO in optinf.h tells whether to project or not.
NTRREP(irrep) tells how many coordinates in irrep which may be of tra-rot character,
 calculated in abatro.F.
NPRREP(irrep) counts how many actually to project out (either 0 or NTRREP(irrep))