hjaaj: peqm control variable should be in gnrinf.h
Sorry Magnus, but have added PEQM to gnrinf.h and removed all "use pe_variables, only: peqm"
A general control variable should not be in a module. If all modules did that the code would overflow
with "use this, only: control1", "use that, only: control2" and so on, and it is not so easy to see what
is general Dalton control and what is internal to a module. If a module should be activated or not is a
general Dalton control.

All tests 'ctest -L peqm' pass after the changes.