hjaaj: fix of walk_vibave* tests (maybe more?)
The essential change is the "NEWGEO = .TRUE." for IWKTYP.eq.6 when
finished (without this, the revised geometry is not transferred to the
slaves for the final calculation during numerical differentiation).

A lot of output improvements triggered by the search for this error.
In particular the output now says "numerical differentiation" and not
"geometry optimization" when doing numerical differentiation with .WALK.