hjaaj: fix of parallel prop_roa (missing NEWGEO) + polish
The problem was that DOCOOR was not transferred to slaves after it was
defined in ABATRO. Sirius had already been called, and NEWGEO was reset
to false after transfer of new geometry. However, transfer of DOCOOR is
also only transferred if NEWGEO true, thus to get right result we need
to reset NEWGEO to true both before and after CALL EXESIR in dalgnr.F
for numerical differentiation.

From my day-long hunt for this error, some improvements in test output.