pno ccsd working with arbitrary cutoffs
PNO CCSD in LSDALTON and DEC implemented. It can be called from the
debug solver right now by using:
.CCSD
.CCDEBUG
.USE_PNOS
in the LSDALTON.INP file. The cutoff threshold is the same for all the
spaces, but this will be changed in the future, once it is clear how to
chose it. It can be specified by
.PNOthr
1.0E-X
and as a debug option one might specify to use the same space for all
occupied indices. this overwrites the previous threshold:
.noPNOtrafo

(input is not case sensitive)