hjaaj: sleeping one whole second was too much for the short test jobs
with, say, 50 fock matrix constructions, the 'call sleep(1)' would add approx 25 seconds
to the original run time, which is a lot if the runtime was a 2 second job.

Now sleep 10 milliseconds on intel ifort, no sleeping with other compilers which do
not have something like the SLEEPQQ of ifort (as far as I know).
One could easily use C binding and call the usleep C routine.
But let us first see how this works with the intel compilers.