hjaaj: save a LOT of time in numerical hessians when C1 symmetry
Hitherto SIRIUS started from scratch after each step because the symmetry will sometimes be broken when symmetry and the whole hessian is calculated.
However! If we calculate with C1 symmetry, there is not symmetry to break, and we can save a lot of time in SIRIUS by letting the wave
function optimization restart.