WARNING I have set NUCDEP = 1000000 hopeing to force an error because clearly it is used uninitialized as clearly stated in the warnings of the dashboard. MB-Grendel-linux-gfortran472-debugMPI. This should be fixed. Also removed "warning: extra tokens at end of #include directive". TK
I encourage all people to look at the warnings. They are there for a reason, and some of them are actual bugs. Like this one. I also suspect that

DaltonGfortranDebugMPI/DALTON/cc/ccsd_energy.F:1481: warning: 'xrtccd' may be used uninitialized in this function

is a bug, but I do not know that code. It is much easier to see if your recent commit have introduced warnings if the list of warning are green! And expected to be green.  TK