Solving further problems on titan + introducing workaround framework
README: I thought it was a good idea to introduce a framework for
workarounds on the different systems we are using, since many of the
problems occur in the kernel/compilers/libraries and we do not usually
have the time to wait for updates. The idea is to keep the original (on
the system at hand failing) code and introduce precompiler flags for
workarounds in the problem causing lines. The workarounds themselves are
to be placed in <daltonroot>/LSDALTON/deccc/dec_workarounds.F90. Please
make sure that the workarounds introduced get a short notice at the
beginning of the dec_workarounds.F90 file such that there is at least
some minimal documentation