hjaaj: fixed a bug for .STEX
With my new cygwin compilation energy_stex failed. After a couple of hours
debugging I discovered the reason was a 'CHARACTER*7 RTNLBL(2)' instead of
'CHARACTER*8 RTNLBL(2)' - looks like a nasty typo.
Nasty because I only discovered the problem now because the CHARACTER*7 caused
accidently another local variable NNDIMX to be modified to a larger integer value,
which later meant a DZERO(WORK(KPRP),NNDIMX) zeroed out of bounds,
which meant that the later MEMREL call caused the Dalton to abort because
the out-of-bounds marker inserted by MEMGET had been destroyed !!!!!!