Since Poul is on the dalton-wizards list I do not send a mail but
in connection with Pouls 70th birthday party (saturday) we decided to make Poul a wikipedia page.

I have now added a first draft to git repo. Clearly in good time ....

LSDALTON/doc_LSDALTON/Poul.text

I need to add some references but I do not know how that is done in wikipedia - Pablo will do that.

So feel free to change the text - we will have a deadline for adding and changing the file that is friday the 28. feb 20.00

best regards
Thomas Kjaergaard