introduced siple timing of phases
the idea is to be able to extract how long a slave worked, communicated
or was idle during a complex subroutine. For that I introduced

call time_start_phase

in the ccsd code. It would of course be cool if other people would use
it, too. So one can get an estimate of how long actual work was done and
how long the slaves just stood idle by taking the difference in the
work/comm/idle times before and after the called subroutine. Currently a
simple/not pecise measure can already be obtained, but the only idle
time there is, is while the slaves wait for a bcast in the mpislave
routine and everything else is considered work (with tehe exception of the
ccsd residual code)