bugfix in get_optimal_batch_sizes_ccsdpt_integrals
i don't know who introduced this (i.e., if it was present before the adapt_to_nnodes logical was introduced or not), but it sure was a tremendous bug. there might be better solutions out there, but this fix at least makes the fullccsdpt test case run through for the MBMR-Sisyphos-linux-gfortran472-omp-int64-openMPI-release. also, inserted some debug prints in the same part of the code to test things across the dashboard.