made original pno scheme work to commit the new file structure
I want to avoid confilcts in the quest for a useful PNO code, which
takes a wile to finish. Therefore I need to merge my stuff into the repo
once in a while. Whilst doing that it turned out that there was an error
in the old code and the test was setup the wrong way. This has been
fixed now.