I got annoyed with the error so I fixed it myself.
I got annoyed with the error, and noone had fixed it in several days so I fixed it myself. Note I do not know what the actual number should be - but at least this way it does not give the compiler error.

/.../DaltonGfortranReleasempich2MPI/DALTON/abacus/dalton.F:471.27:

         read(env_value, '(i)') omp_num_threads
                           1

Error: Nonnegative width required in format string at (1)