Changed testcase decmp2_RI
The decmp2_RI testcase has been very unstable wrt to
changes in the orbital localization procedure since it tested for
quite many decimals in the correlation energy while a FOT value of
only 2 was used. I lowered the FOT by setting FOT value to 3, and
test only decimal places which makes sense compared to the error
being made in the correlation energy as a result of the given
FOT value.