Moved prop_* test cases to runtest.
prop_roa will fail in parallel. The reason is that with DIRECT,
the three first time (of 15 times) CID is calculated it is not
identical to the numbers obtained with NON-DIRECT.

prop_vibana is no longer failing, but only because I introduced
"MAXRED too small in ANRCTL" as a valid error message.