Added warning for Bohr vs Angstrom.
Cleanup of the "recommended output" printout - to include anions (thanks to hjaaj)
Added calculation of Short Hydrogen-X bonds and other Y-X bonds and number of bonds
so as to warn in case people gave input in Angstrom but forgot the Angstrom keyword, or
the other way around. It is not bulletproof but it will hopefully help.

TK