Added a WARNING when the SCF gradient is larger than the FOT
In this case it is very likely that the SCF error (error in MO coef) would
dominate the error in the DEC calculation. For now I have only implemented it
in the dec_main_prog_input since I did not fully know where to put it.
Please someone move it!

And I also made a subroutine get_AO_gradientnormFull(F, D, S, gradnorm,nbast)
which takes square fortran arrays as input.

TK