Localization of few orbitals
Made changes so localization procedure is run even if
you only have two orbitals. The change checks the number of
orbitals and skips starting trialvectors in the davidson
solver to avoid troubles with linear dependencies of the
trial vectors. Also, make sure that routine is not exited in
first macro iteration to prevent exiting when starting from a
symmetry minimum.