Modified the XC code. I changed the memory handling related to OpenMP.
Before each OpenMP thread made its own version of the output
and added to it through the XC integration.
This avoided the need for CRITICAL statements. However for big systems the
added memory required is not desireable. This was therefore reverted and the
timings are basicly unaffected with the current scheme using the ATOMIC clause.

 TK