Added some comment line to a error statement.
Kasper had problems with the EMSL format. So added text so that
the user is now told that a possible reason is that the line in the
basisset file is too long and that it must be split.

I was unable to reproduce the strange problem Kasper had concerning "space" vs. "no space" in the second line.

TK