hjaaj: fix of energy_solv, mcscf, and mcscf_nomp2 tests
Now again we skip continuum models (old spherical cavity and PCM) in RHF and MP2
if we continue with MCSCF or CI.
energy_solv failed because the HF and MP2 energies in the test were not updated to the new
energies including solvent contribution. However, gone back, solvent just takes extra time here.
mcscf and mcscf_nomp2 - which really are pcm tests and ought to have had pcm in the names (!) -
failed because of a bug in the PCM code: if you include PCM in both RHF and MCSCF, then the
PCM contibution in MCSCF is totally wrong ! PCM people ought to look into that. However,
things work again because I now have disabled PCM again in RHF if followed by MCSCF.

Added "mcscf" to essential tests.