Modified the code so that the XC code can be called from DEC.
Before there was a problem when a local masters tried to calc a XC
contribution because the local master had a different number of nodes
than was originally used to build the grid. Now the grid is
reconstructed everytime it is called from DEC - this can be modified
but for now we want to be safe. TK