Added a test on MPI memory leak!
I modified LSDALTON_Link to include the test

f.add(string = 'Allocated MPI memory a cross all slaves')

In case of non MPI, the output file will read

  Allocated MPI memory a cross all slaves:          0 byte  - Should be zero - otherwise a leakage is present
  This is a non MPI calculation so naturally no memory is allocated on slaves!

In case of MPI, the output file will read

  Allocated MPI memory a cross all slaves:          0 byte  - Should be zero - otherwise a leakage is present

IMPORTANT Note that this print out only makes sense with the keyword

**INFO
.DEBUG_MPI_MEM

If this keyword is left out the memory usage statistics are not communicated to the master and the output file will
not contain this printout.

I suggest to use this keyword and filter option on all testcases!

TK