Implemented the ability to chose a subset of the MPI processes to be included in the Scalapack format.
The memory of the matrices will only be distributed among these MPI processes
And only these MPI processes will be included in the paralelization of the
matrix operations. This have become important since a DEC MP2 geometry
optimization have significantly different MPI scaling behavior for the matrix
operations, the integrals and the DEC part.

TK