fixing MP2 testcase and removing MODEL_CCSDpT from solver
I removed the MODEL_CCSDpT from the solver since this is not the
wavefunction parameters that are requested from the solvers point of
view. I know it is a bit "cumbersome" but it requires a local
solver_ccmodel argument that is set to MODEL_CCSD if MODEL_CCSDpT is
requested. I think that makes the code more obvious and avoids
references to a model the solver cannot treat within the ccsolver_par
routine