New default and keywords for fragment optimization
.FRAG_EXP_SIZE new default is 5

Now the reduction is done in both the occupied and
the virtual space at the same time.

To start reducing the occupied space first set:
.FRAG_RED_OCC

To start reducing the virtual space first set:
.FRAG_RED_VIRT

The Threshold use for the convergence of the first
space is set to 0.9*FOT by default and can be controled via:
.FRAG_RED1_THR
 X

where X .le. 1 and the threshold becomes X*FOT

In the same way for the second space to be reduced,
.FRAG_RED2_THR
 Y

Again, by default both spaces are reduced at the same
time and the threshold use is thus the FOT.