Fixed memory leak in F12 code. Added a Integralthreshold for DEC.
Before the same integral threshold was used for HF and DEC. Which does not always make sense.

We automaticly tighten the integral threshold when using .CONVDYN TIGHT which is often done for DEC. However there is no reason to use a integral threshold of 10.0E-11_realk when the FOT is 1.0E-4_realk. TK

For now the default integral threshold in DEC is chosen to be 10.0E-10_realk
But I might set this in relation to FOT.
Maybe IntegralThreshold = 1.0E-4_realk*FOT.

TK