Change the Format of the dens.restart, lcv_basis.restart.
Before the dens.restart, lcv_basis.restart included a logical with a size decided by the compiler.
Some compilers use 64 bit logicals when the user sepcifies 64 bit integers.
Other compilers use 32 bit logicals when the user sepcifies 64 bit integers.
Now a 64 bit logical is hardcoded and the format should now be independent of the compilation of the program.

TK