hjaaj: extended prop_vibg2 test
with test of that it works when you ask for more excitations than
number of variables.

ALSO appended the old description from prop_vibg2.info to the prop_vibg2.dal
file. THIS SHOULD BE DONE WITH ALL THE OLD .info files in my opinion.
Often I cannot guess the whole purpose with the specific tests without this
information, so valuable information is lost by not including this information
for the developers.