Fixed STUPID mistake. Someone should be punched hard in the face!
Please Understand that you are NOT ALLOWED to
use a optional argument unless you have tested
that the argument is present.

in the case of an if statment like

if(present(dft).AND.dft)then

you cannot assume that the compiler first test the first statement "present(dft)"
and then only if this is true test the second statement.

This is at the prerogative of the compiler!