Added a Keyword .AUXATOMICEXTENT
This keyword makes it possible to use the full molecular system for RI.

The default right now is to use the Aux AO functions on all atoms included in the
Atomic extent. In principle the RI summation should run over all Aux AOs in the molecule, but
due to the local nature it should be possible to restrict the Aux AO sum (the default).

This keyword is mainly to test that RI restriction approximation. TK