Added the CABSP basis set reading.
CABSP=aug-cc-pVDZ will add the basis set aug-cc-pVDZ to
the regular basis set and ensure that the regular basis is a
subset of the CABS basis. Note that CABSP is read and added with REGULAR
and stored in CABS. So CABSP and CABS cannot be used together.
added testcase. TK