hjaaj: one more attempt to fix a single open shell ROHF
The problem is that we have two different SCF methods for ROHF doublet,
the original activated with .OPEN SHELL and the HSROHF code.
Some of the property modules expect one form and some the other.
Fixing one caused the other to fail, as you may have noticed.
The problems are generally caused by that the two methods save
different FC and FV matrices on the interface file SIRIFC.
I have now cleaned up the code, saved on SIRIFC with MCTYPE=-1 if
HSROHF, and then in RESPONS recalculate the FC and FV if RESPONS
code uses HSROHF code but non-HSROHF FC,FV saved on SIRIFC.

And HSFOCK in scbrhf.h eliminated, it didn't help to have both
HSFOCK in scbrhf.h and HSROHF in infinp.h basically meaning the same thing.