fix MB-Grendel ... debugMPI ccsd fail
stupid compiler, but still necessary to adapt

obviously the compiler had problems interpreting an abstract interface
used in a procedure pointer as the prodedure. That means it was possible
to call the function but it returned NANs. Now I have declared the procedure
pointer without an abstract interface and it works ... we should really
consider removing such compilers from the build system.