introduced simple estimate in frag energies and reduced printout
for large jobs (>100 nodes) the printout will now be reduced unless the
print level is set to >1. At some point we should really streamline all
the print levels. I made a suggestion how that may be done in
dec_settings.F90 where the print level default is set

for the most expanded fragment we can now use the energy of the most
expanded fragment minus the energy of the converged fragment