add .NO_PAIRS keyword to skip pair calculations
This is a debuging keyword so by default the
pair estimate calculations are still performed.

After a calculation with the no_pairs keyword
the restart option should work and read the
optimized fragment as well as the pair estimates
and then only perform pair calculations if the
no_pairs keyword has been removed.