workaround for compiler assoicated problem
the gnu 4.7 probably tried to allocate an array with an unitialized
value on the stack and therefore ran out of memory. this was related to
the way the arguments were passed to a subroutine, i.e. as array instead
of indiviudal integers. Now, passing the integers individually, the routine
works