Optimized the MPI build of Calpha in RIMP2. TK
It may not sound like a optimization but instead of using all "nodtot" nodes
to calculate Calpha, I use a subset of "nodtot". The reason is that Calpha is
dominated by MPI communication and the more nodes the more MPI communication.
Building Calpha is a small part of a RIMP2 calculation, and looking at the
scaling curve for this contribution it makes sense to simply reduce the
number of nodes involved in this part of the calculation.