Second atempt to test the output parser
The output_plot.py script had to be
modify to use a non-interacting backend
to display plots (figure have to be saved in files)

In decmp2_energy the dec parser is used as well as
the plot of pair interaction energy.

In fullccsdpt_print, the cc parser and the dec parser
for full cc jobs are used.