Added a keyword .DENSELEVEL2 which performs the second level in a TRILEVEL starting guess using a DENSE matrix type.
Due to the vastly different number of basis functions in Level 2 and Level 3
it makes sense to treat these using different methods.

Also Modified the way the dens.restart is saved. Before the level 2 density was saved as
dens.restart and if the calculation crashed during level 2 you had to manually rename it to
vdens.restart in order to restart the calculation. The dens.restart was then renamed by the program
to vdens.restart before starting the level 3 SCF procedure.
Now the SCF algorithm saves the density matrix as vdens.restart
during level 2 and saves it as dens.restart during level 3.

This means you do not have to manual rename. Simply restart the calc. TK