Moved the Background buffer outside the fragment calculation.
So now the background buffer is done only twice.

once before all the atomic fragment optimization/pair estimates and
once before all the pair jobs.

And the size is determined by a keyword .BG_MEMORY (in GigaBytes) and
it is currently unclear how big this should be, but it should basicly be
.MEMORY - memory used for EOS amplitudes and integrals - so like 60-90 %
depending on the FOT. For Pair estimates and Pairs this memory usage can
be calculated on runtime, but not for atomic fragment optimization. TK