hjaaj: fix of mpi problem with solution reported by Arseny Slobodyuk on Dalton forum
From the forum:
I had the same errors on Debian Linux Wheezy and Jessie when configured to use MPI (./setup --mpi).
Sorry about necroposting, but this thread seems to be the only one about the problem.
My config: MPI: openmpi 1.4.5; compiler: mpif90; GNU BLAS & LAPACK
I've solved it with this patch. IDK exactly how to check if it is really correct.