hjaaj: not very robust test; new output file saved
My tests showed that quite small changes in DFT convergence
can make the number of geometry iterations change, presumably
because of the very flat vdW surface. The test would be
more robust if it just tested on final geometry and not
on the number of geometry iterations, but I will leave the
decision on any such change to those who made this D2 test.