add cmake setup flag ENABLE_LARGE_TEST
This will allow us to perform weekly testing
that should be able to test more real-life
large scale features (see large_decrimp2)

This flag is also used to change the ctest timeout.
It should not be used in the traditional dash scripts
as those test need to be run using a queueing system.