adding hercules testcases such that everybody can run them
I added the run.bash script and use a folder structure, such that it is
simple for everybody to run the testases on a given machine. The idea
is, that testing can be invoked by running

run.bash <machinename> <nightly/continuous> <path/to/tmp/dir>

which is realized by usint the folder structure
cdash/machinename/nighly
cdash/machinename/continuous

and the run.bash will run all the .bash files in the given folders. In
this way everybody who has access to a given machine can simply add
run.bash to their crontabs