IMPORTANT!!!
Added a global logical "developer" in ls_parameters.F90 in an attempt to avoid that the general users tries out bad keyword combinations.

An example. For the next Dalton version I would like to release SNOOP in combination with **CC since this has been properly tested. SNOOP in combination with **DEC can in principle be calculated, but it has not been tested, and the best strategy is still unclear. Therefore I don't want the general user to be able to run SNOOP/DEC calculations. But I myself would like to be able to do this for further testing. So I would like a solution which protects the general user from DEC/SNOOP but which at the same time does not prevent me from doing more testing of SNOOP/DEC.
I therefore made a global logical "developer" in ls_parameters, which is false by default. But you can make true e.g. by setting:

**GENERAL
.DEVELOPER

Inside DEC input check I then have:

   if(DECinfo%SNOOP) then

       if(.not. DECinfo%full_molecular_cc) then
          if(.not. developer) then
             call lsquit('SNOOP and DEC are not compatible yet!',-1)
          end if
       end if

  end if

Then, the general user will hit the quit statement and not be able to run SNOOP/DEC calculations. I, as a developer, can overrule this by setting the **GENERAL/.DEVELOPER.

Thus, anyone who uses the **GENERAL/.DEVELOPER should be aware of the fact that the code they use is not well-tested. So only people who "know what they are doing" should know of this keyword. Of course, if the general user finds out about this keyword and uses it, then he/she is on his own, and we cannot be held responsible for faulty results.