BugFix: FC changed to FCS when required in DFTKSMGGASPIN routine + introduce IPNT in DFTKSM + DFTKSMGGASPIN calls (for debugging purposes)
Important change in SRDFT1JT routine:
"if (domcsrdft .and. docisrdft)" is true in OPST routine i.e.
here we should *not* use spin! This special case has been inserted,
and works for LSDA (and GGA without spin)
- yet there remains things to be changed: the arrays d1Ex (and d1Ec) will
come in wrong for this special case if a GGA SPIN functional is specificed
(this is not fixed in this commit!)

Other changes: Debug write statements have been inserted